Clean up

+++ Tests/MapSimple.cpp
@@ -64,30 +64,30 @@ void testFull(Func func)
 void tsImplCreation(auto map)
 {
 
-	EXPECT_EQ(map.getLen(), 0);
+	EXPECT_EQ(map.getLen(), size_t(0));
 
 	for (size_t numOfIters{ 0 }; numOfIters < 2; ++numOfIters)
 	{
 
 		map[0] = 101;
-		EXPECT_EQ(map.getLen(), 1);
-		EXPECT_EQ(map[0], 101);
+		EXPECT_EQ(map.getLen(), size_t(1));
+		EXPECT_EQ(map[0], size_t(101));
 		map[1] = 102;
-		EXPECT_EQ(map.getLen(), 2);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
+		EXPECT_EQ(map.getLen(), size_t(2));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
 
 		map.insert(2, 103);
-		EXPECT_EQ(map.getLen(), 3);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(3));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map[1] += 100;
-		EXPECT_EQ(map.getLen(), 3);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 202);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(3));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(202));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map.clear();
 
@@ -109,30 +109,30 @@ void tsImplRemove(auto map)
 	{
 
 		map[0] = 101;
-		EXPECT_EQ(map.getLen(), 1);
-		EXPECT_EQ(map[0], 101);
+		EXPECT_EQ(map.getLen(), size_t(1));
+		EXPECT_EQ(map[0], size_t(101));
 		map[1] = 102;
-		EXPECT_EQ(map.getLen(), 2);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
+		EXPECT_EQ(map.getLen(), size_t(2));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
 
 		map.insert(2, 103);
-		EXPECT_EQ(map.getLen(), 3);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(3));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map.remove(1);
-		EXPECT_EQ(map.getLen(), 2);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(2));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map.remove(0);
-		EXPECT_EQ(map.getLen(), 1);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(1));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map.remove(2);
-		EXPECT_EQ(map.getLen(), 0);
+		EXPECT_EQ(map.getLen(), size_t(0));
 
 		map.remove(2);
 		map.remove(1);
@@ -141,18 +141,18 @@ void tsImplRemove(auto map)
 		map.clear();
 
 		map[0] = 101;
-		EXPECT_EQ(map.getLen(), 1);
-		EXPECT_EQ(map[0], 101);
+		EXPECT_EQ(map.getLen(), size_t(1));
+		EXPECT_EQ(map[0], size_t(101));
 		map[1] = 102;
-		EXPECT_EQ(map.getLen(), 2);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
+		EXPECT_EQ(map.getLen(), size_t(2));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
 
 		map.insert(2, 103);
-		EXPECT_EQ(map.getLen(), 3);
-		EXPECT_EQ(map[0], 101);
-		EXPECT_EQ(map[1], 102);
-		EXPECT_EQ(map[2], 103);
+		EXPECT_EQ(map.getLen(), size_t(3));
+		EXPECT_EQ(map[0], size_t(101));
+		EXPECT_EQ(map[1], size_t(102));
+		EXPECT_EQ(map[2], size_t(103));
 
 		map.clear();
 
@@ -176,16 +176,16 @@ void tsImplFind(auto map)
 		EXPECT_FALSE(map.find(0).has_value());
 
 		map[0] = 101;
-		EXPECT_EQ(map.find(0).value(), 101);
+		EXPECT_EQ(map.find(0).value(), size_t(101));
 		EXPECT_FALSE(map.find(1).has_value());
 
 		map[1] = 102;
-		EXPECT_EQ(map.find(0).value(), 101);
-		EXPECT_EQ(map.find(1).value(), 102);
+		EXPECT_EQ(map.find(0).value(), size_t(101));
+		EXPECT_EQ(map.find(1).value(), size_t(102));
 
 		map.remove(0);
 		EXPECT_FALSE(map.find(0).has_value());
-		EXPECT_EQ(map.find(1).value(), 102);
+		EXPECT_EQ(map.find(1).value(), size_t(102));
 
 		map.clear();
 		EXPECT_FALSE(map.find(0).has_value());
@@ -237,7 +237,7 @@ void tsImplForEachItem(auto map)
 	map[30] = 103;
 
 	const auto& items{ map.getItems() };
-	int expectedItem = 101;
+	size_t expectedItem = 101;
 
 	for (auto item : items)
 	{
@@ -245,7 +245,7 @@ void tsImplForEachItem(auto map)
 		++expectedItem;
 	}
 
-	EXPECT_EQ(expectedItem, 104);
+	EXPECT_EQ(expectedItem, size_t(104));
 
 }
 
@@ -264,7 +264,7 @@ void tsImplForEach(auto map)
 	map[30] = 103;
 
 	int expectedKey = 10;
-	int expectedItem = 101;
+	size_t expectedItem = 101;
 
 	for (const auto& [key, item] : map)
 	{
@@ -275,7 +275,7 @@ void tsImplForEach(auto map)
 	}
 
 	EXPECT_EQ(expectedKey, 40);
-	EXPECT_EQ(expectedItem, 104);
+	EXPECT_EQ(expectedItem, size_t(104));
 
 }
